refactor: remove article service
