+++ src/appv2/features/article/mod.rs
@@ -3,5 +3,4 @@ pub mod entities;
 pub mod presenters;
 pub mod repositories;
 pub mod requests;
-pub mod services;
 pub mod usecases;
